add-hooks-deprecation-test
