+++ pytorch_lightning/core/datamodule.py
@@ -21,7 +21,7 @@ from torch.utils.data import DataLoader, Dataset, IterableDataset
 
 from pytorch_lightning.core.hooks import CheckpointHooks, DataHooks
 from pytorch_lightning.utilities.argparse import add_argparse_args, from_argparse_args, get_init_arguments_and_types
-from pytorch_lightning.utilities.distributed import rank_zero_deprecation, rank_zero_only, rank_zero_warn
+from pytorch_lightning.utilities.distributed import rank_zero_deprecation, rank_zero_only
 
 
 class LightningDataModule(CheckpointHooks, DataHooks):
@@ -434,7 +434,7 @@ class LightningDataModule(CheckpointHooks, DataHooks):
                 obj._has_prepared_data = True
 
             if has_run:
-                rank_zero_warn(
+                rank_zero_deprecation(
                     f"DataModule.{name} has already been called, so it will not be called again. "
                     f"In v1.6 this behavior will change to always call DataModule.{name}."
                 )
