+++ tests/deprecated_api/test_remove_1-6.py
@@ -108,3 +108,54 @@ def test_v1_6_0_datamodule_lifecycle_properties(tmpdir):
         dm.has_teardown_test
     with pytest.deprecated_call(match=r"DataModule property `has_teardown_predict` was deprecated in v1.4"):
         dm.has_teardown_predict
+
+
+def test_v1_6_0_datamodule_hooks_calls(tmpdir):
+    """Test that repeated calls to DataHooks' hooks show a warning about the coming API change."""
+
+    class TestDataModule(BoringDataModule):
+        setup_calls = []
+        teardown_calls = []
+        prepare_data_calls = 0
+
+        def setup(self, stage=None):
+            super().setup(stage=stage)
+            self.setup_calls.append(stage)
+
+        def teardown(self, stage=None):
+            super().teardown(stage=stage)
+            self.teardown_calls.append(stage)
+
+        def prepare_data(self):
+            super().prepare_data()
+            self.prepare_data_calls += 1
+
+    dm = TestDataModule()
+    dm.prepare_data()
+    dm.prepare_data()
+    dm.setup('fit')
+    with pytest.deprecated_call(
+        match=r"DataModule.setup has already been called, so it will not be called again. "
+        "In v1.6 this behavior will change to always call DataModule.setup"
+    ):
+        dm.setup('fit')
+    dm.setup()
+    dm.setup()
+    dm.teardown('validate')
+    with pytest.deprecated_call(
+        match=r"DataModule.teardown has already been called, so it will not be called again. "
+        "In v1.6 this behavior will change to always call DataModule.teardown"
+    ):
+        dm.teardown('validate')
+
+    assert dm.prepare_data_calls == 1
+    assert dm.setup_calls == ['fit', None]
+    assert dm.teardown_calls == ['validate']
+
+    trainer = Trainer(default_root_dir=tmpdir, fast_dev_run=1)
+    trainer.test(BoringModel(), datamodule=dm)
+
+    # same number of calls
+    assert dm.prepare_data_calls == 1
+    assert dm.setup_calls == ['fit', None]
+    assert dm.teardown_calls == ['validate', 'test']
